Recipe detail screen fix

+++ expire_app/lib/models/recipe_details.dart
@@ -1,8 +1,3 @@
-// To parse this JSON data, do
-//
-//     final recipeDetails = recipeDetailsFromJson(jsonString);
-
-import 'package:meta/meta.dart';
 import 'dart:convert';
 
 RecipeDetails recipeDetailsFromJson(String str) => RecipeDetails.fromJson(json.decode(str));
@@ -61,8 +56,8 @@ class RecipeDetails {
   final String gaps;
   final bool lowFodmap;
   final int aggregateLikes;
-  final int spoonacularScore;
-  final int healthScore;
+  final double spoonacularScore;
+  final double healthScore;
   final String creditsText;
   final String license;
   final String sourceName;
@@ -87,84 +82,84 @@ class RecipeDetails {
   final String spoonacularSourceUrl;
 
   factory RecipeDetails.fromJson(Map<String, dynamic> json) => RecipeDetails(
-        vegetarian: json["vegetarian"],
-        vegan: json["vegan"],
-        glutenFree: json["glutenFree"],
-        dairyFree: json["dairyFree"],
-        veryHealthy: json["veryHealthy"],
-        cheap: json["cheap"],
-        veryPopular: json["veryPopular"],
-        sustainable: json["sustainable"],
-        weightWatcherSmartPoints: json["weightWatcherSmartPoints"],
-        gaps: json["gaps"],
-        lowFodmap: json["lowFodmap"],
-        aggregateLikes: json["aggregateLikes"],
-        spoonacularScore: json["spoonacularScore"],
-        healthScore: json["healthScore"],
-        creditsText: json["creditsText"],
-        license: json["license"],
-        sourceName: json["sourceName"],
-        pricePerServing: json["pricePerServing"].toDouble(),
-        extendedIngredients:
-            List<ExtendedIngredient>.from(json["extendedIngredients"].map((x) => ExtendedIngredient.fromJson(x))),
-        id: json["id"],
-        title: json["title"],
-        readyInMinutes: json["readyInMinutes"],
-        servings: json["servings"],
-        sourceUrl: json["sourceUrl"],
-        image: json["image"],
-        imageType: json["imageType"],
-        summary: json["summary"],
-        cuisines: List<dynamic>.from(json["cuisines"].map((x) => x)),
-        dishTypes: List<String>.from(json["dishTypes"].map((x) => x)),
-        diets: List<String>.from(json["diets"].map((x) => x)),
-        occasions: List<dynamic>.from(json["occasions"].map((x) => x)),
-        winePairing: WinePairing.fromJson(json["winePairing"]),
-        instructions: json["instructions"],
-        analyzedInstructions:
-            List<AnalyzedInstruction>.from(json["analyzedInstructions"].map((x) => AnalyzedInstruction.fromJson(x))),
-        originalId: json["originalId"],
-        spoonacularSourceUrl: json["spoonacularSourceUrl"],
-      );
+    vegetarian: json["vegetarian"],
+    vegan: json["vegan"],
+    glutenFree: json["glutenFree"],
+    dairyFree: json["dairyFree"],
+    veryHealthy: json["veryHealthy"],
+    cheap: json["cheap"],
+    veryPopular: json["veryPopular"],
+    sustainable: json["sustainable"],
+    weightWatcherSmartPoints: json["weightWatcherSmartPoints"],
+    gaps: json["gaps"],
+    lowFodmap: json["lowFodmap"],
+    aggregateLikes: json["aggregateLikes"],
+    spoonacularScore: json["spoonacularScore"],
+    healthScore: json["healthScore"],
+    creditsText: json["creditsText"],
+    license: json["license"],
+    sourceName: json["sourceName"],
+    pricePerServing: json["pricePerServing"].toDouble(),
+    extendedIngredients:
+    List<ExtendedIngredient>.from(json["extendedIngredients"].map((x) => ExtendedIngredient.fromJson(x))),
+    id: json["id"],
+    title: json["title"],
+    readyInMinutes: json["readyInMinutes"],
+    servings: json["servings"],
+    sourceUrl: json["sourceUrl"],
+    image: json["image"],
+    imageType: json["imageType"],
+    summary: json["summary"],
+    cuisines: List<dynamic>.from(json["cuisines"].map((x) => x)),
+    dishTypes: List<String>.from(json["dishTypes"].map((x) => x)),
+    diets: List<String>.from(json["diets"].map((x) => x)),
+    occasions: List<dynamic>.from(json["occasions"].map((x) => x)),
+    winePairing: WinePairing.fromJson(json["winePairing"]),
+    instructions: json["instructions"],
+    analyzedInstructions:
+    List<AnalyzedInstruction>.from(json["analyzedInstructions"].map((x) => AnalyzedInstruction.fromJson(x))),
+    originalId: json["originalId"],
+    spoonacularSourceUrl: json["spoonacularSourceUrl"],
+  );
 
   Map<String, dynamic> toJson() => {
-        "vegetarian": vegetarian,
-        "vegan": vegan,
-        "glutenFree": glutenFree,
-        "dairyFree": dairyFree,
-        "veryHealthy": veryHealthy,
-        "cheap": cheap,
-        "veryPopular": veryPopular,
-        "sustainable": sustainable,
-        "weightWatcherSmartPoints": weightWatcherSmartPoints,
-        "gaps": gaps,
-        "lowFodmap": lowFodmap,
-        "aggregateLikes": aggregateLikes,
-        "spoonacularScore": spoonacularScore,
-        "healthScore": healthScore,
-        "creditsText": creditsText,
-        "license": license,
-        "sourceName": sourceName,
-        "pricePerServing": pricePerServing,
-        "extendedIngredients": List<dynamic>.from(extendedIngredients.map((x) => x.toJson())),
-        "id": id,
-        "title": title,
-        "readyInMinutes": readyInMinutes,
-        "servings": servings,
-        "sourceUrl": sourceUrl,
-        "image": image,
-        "imageType": imageType,
-        "summary": summary,
-        "cuisines": List<dynamic>.from(cuisines.map((x) => x)),
-        "dishTypes": List<dynamic>.from(dishTypes.map((x) => x)),
-        "diets": List<dynamic>.from(diets.map((x) => x)),
-        "occasions": List<dynamic>.from(occasions.map((x) => x)),
-        "winePairing": winePairing.toJson(),
-        "instructions": instructions,
-        "analyzedInstructions": List<dynamic>.from(analyzedInstructions.map((x) => x.toJson())),
-        "originalId": originalId,
-        "spoonacularSourceUrl": spoonacularSourceUrl,
-      };
+    "vegetarian": vegetarian,
+    "vegan": vegan,
+    "glutenFree": glutenFree,
+    "dairyFree": dairyFree,
+    "veryHealthy": veryHealthy,
+    "cheap": cheap,
+    "veryPopular": veryPopular,
+    "sustainable": sustainable,
+    "weightWatcherSmartPoints": weightWatcherSmartPoints,
+    "gaps": gaps,
+    "lowFodmap": lowFodmap,
+    "aggregateLikes": aggregateLikes,
+    "spoonacularScore": spoonacularScore,
+    "healthScore": healthScore,
+    "creditsText": creditsText,
+    "license": license,
+    "sourceName": sourceName,
+    "pricePerServing": pricePerServing,
+    "extendedIngredients": List<dynamic>.from(extendedIngredients.map((x) => x.toJson())),
+    "id": id,
+    "title": title,
+    "readyInMinutes": readyInMinutes,
+    "servings": servings,
+    "sourceUrl": sourceUrl,
+    "image": image,
+    "imageType": imageType,
+    "summary": summary,
+    "cuisines": List<dynamic>.from(cuisines.map((x) => x)),
+    "dishTypes": List<dynamic>.from(dishTypes.map((x) => x)),
+    "diets": List<dynamic>.from(diets.map((x) => x)),
+    "occasions": List<dynamic>.from(occasions.map((x) => x)),
+    "winePairing": winePairing.toJson(),
+    "instructions": instructions,
+    "analyzedInstructions": List<dynamic>.from(analyzedInstructions.map((x) => x.toJson())),
+    "originalId": originalId,
+    "spoonacularSourceUrl": spoonacularSourceUrl,
+  };
 }
 
 class AnalyzedInstruction {
@@ -177,14 +172,14 @@ class AnalyzedInstruction {
   final List<Step> steps;
 
   factory AnalyzedInstruction.fromJson(Map<String, dynamic> json) => AnalyzedInstruction(
-        name: json["name"],
-        steps: List<Step>.from(json["steps"].map((x) => Step.fromJson(x))),
-      );
+    name: json["name"],
+    steps: List<Step>.from(json["steps"].map((x) => Step.fromJson(x))),
+  );
 
   Map<String, dynamic> toJson() => {
-        "name": name,
-        "steps": List<dynamic>.from(steps.map((x) => x.toJson())),
-      };
+    "name": name,
+    "steps": List<dynamic>.from(steps.map((x) => x.toJson())),
+  };
 }
 
 class Step {
@@ -203,20 +198,20 @@ class Step {
   final Length? length;
 
   factory Step.fromJson(Map<String, dynamic> json) => Step(
-        number: json["number"],
-        step: json["step"],
-        ingredients: List<Ent>.from(json["ingredients"].map((x) => Ent.fromJson(x))),
-        equipment: List<Ent>.from(json["equipment"].map((x) => Ent.fromJson(x))),
-        length: json["length"] == null ? null : Length.fromJson(json["length"]),
-      );
+    number: json["number"],
+    step: json["step"],
+    ingredients: List<Ent>.from(json["ingredients"].map((x) => Ent.fromJson(x))),
+    equipment: List<Ent>.from(json["equipment"].map((x) => Ent.fromJson(x))),
+    length: json["length"] == null ? null : Length.fromJson(json["length"]),
+  );
 
   Map<String, dynamic> toJson() => {
-        "number": number,
-        "step": step,
-        "ingredients": List<dynamic>.from(ingredients.map((x) => x.toJson())),
-        "equipment": List<dynamic>.from(equipment.map((x) => x.toJson())),
-        "length": length == null ? null : length!.toJson(),
-      };
+    "number": number,
+    "step": step,
+    "ingredients": List<dynamic>.from(ingredients.map((x) => x.toJson())),
+    "equipment": List<dynamic>.from(equipment.map((x) => x.toJson())),
+    "length": length == null ? null : length!.toJson(),
+  };
 }
 
 class Ent {
@@ -233,18 +228,18 @@ class Ent {
   final String image;
 
   factory Ent.fromJson(Map<String, dynamic> json) => Ent(
-        id: json["id"],
-        name: json["name"],
-        localizedName: json["localizedName"],
-        image: json["image"],
-      );
+    id: json["id"],
+    name: json["name"],
+    localizedName: json["localizedName"],
+    image: json["image"],
+  );
 
   Map<String, dynamic> toJson() => {
-        "id": id,
-        "name": name,
-        "localizedName": localizedName,
-        "image": image,
-      };
+    "id": id,
+    "name": name,
+    "localizedName": localizedName,
+    "image": image,
+  };
 }
 
 class Length {
@@ -257,14 +252,14 @@ class Length {
   final String unit;
 
   factory Length.fromJson(Map<String, dynamic> json) => Length(
-        number: json["number"],
-        unit: json["unit"],
-      );
+    number: json["number"],
+    unit: json["unit"],
+  );
 
   Map<String, dynamic> toJson() => {
-        "number": number,
-        "unit": unit,
-      };
+    "number": number,
+    "unit": unit,
+  };
 }
 
 class ExtendedIngredient {
@@ -294,45 +289,45 @@ class ExtendedIngredient {
   final String original;
   final String originalString;
   final String originalName;
-  final int amount;
+  final double amount;
   final String unit;
   final List<String> meta;
   final List<String> metaInformation;
   final Measures measures;
 
   factory ExtendedIngredient.fromJson(Map<String, dynamic> json) => ExtendedIngredient(
-        id: json["id"],
-        aisle: json["aisle"],
-        image: json["image"],
-        consistency: json["consistency"],
-        name: json["name"],
-        nameClean: json["nameClean"],
-        original: json["original"],
-        originalString: json["originalString"],
-        originalName: json["originalName"],
-        amount: json["amount"],
-        unit: json["unit"],
-        meta: List<String>.from(json["meta"].map((x) => x)),
-        metaInformation: List<String>.from(json["metaInformation"].map((x) => x)),
-        measures: Measures.fromJson(json["measures"]),
-      );
+    id: json["id"],
+    aisle: json["aisle"],
+    image: json["image"],
+    consistency: json["consistency"],
+    name: json["name"],
+    nameClean: json["nameClean"],
+    original: json["original"],
+    originalString: json["originalString"],
+    originalName: json["originalName"],
+    amount: json["amount"],
+    unit: json["unit"],
+    meta: List<String>.from(json["meta"].map((x) => x)),
+    metaInformation: List<String>.from(json["metaInformation"].map((x) => x)),
+    measures: Measures.fromJson(json["measures"]),
+  );
 
   Map<String, dynamic> toJson() => {
-        "id": id,
-        "aisle": aisle,
-        "image": image,
-        "consistency": consistency,
-        "name": name,
-        "nameClean": nameClean,
-        "original": original,
-        "originalString": originalString,
-        "originalName": originalName,
-        "amount": amount,
-        "unit": unit,
-        "meta": List<dynamic>.from(meta.map((x) => x)),
-        "metaInformation": List<dynamic>.from(metaInformation.map((x) => x)),
-        "measures": measures.toJson(),
-      };
+    "id": id,
+    "aisle": aisle,
+    "image": image,
+    "consistency": consistency,
+    "name": name,
+    "nameClean": nameClean,
+    "original": original,
+    "originalString": originalString,
+    "originalName": originalName,
+    "amount": amount,
+    "unit": unit,
+    "meta": List<dynamic>.from(meta.map((x) => x)),
+    "metaInformation": List<dynamic>.from(metaInformation.map((x) => x)),
+    "measures": measures.toJson(),
+  };
 }
 
 class Measures {
@@ -345,14 +340,14 @@ class Measures {
   final Metric metric;
 
   factory Measures.fromJson(Map<String, dynamic> json) => Measures(
-        us: Metric.fromJson(json["us"]),
-        metric: Metric.fromJson(json["metric"]),
-      );
+    us: Metric.fromJson(json["us"]),
+    metric: Metric.fromJson(json["metric"]),
+  );
 
   Map<String, dynamic> toJson() => {
-        "us": us.toJson(),
-        "metric": metric.toJson(),
-      };
+    "us": us.toJson(),
+    "metric": metric.toJson(),
+  };
 }
 
 class Metric {
@@ -362,21 +357,21 @@ class Metric {
     required this.unitLong,
   });
 
-  final int amount;
+  final double amount;
   final String unitShort;
   final String unitLong;
 
   factory Metric.fromJson(Map<String, dynamic> json) => Metric(
-        amount: json["amount"],
-        unitShort: json["unitShort"],
-        unitLong: json["unitLong"],
-      );
+    amount: json["amount"],
+    unitShort: json["unitShort"],
+    unitLong: json["unitLong"],
+  );
 
   Map<String, dynamic> toJson() => {
-        "amount": amount,
-        "unitShort": unitShort,
-        "unitLong": unitLong,
-      };
+    "amount": amount,
+    "unitShort": unitShort,
+    "unitLong": unitLong,
+  };
 }
 
 class WinePairing {
